+++ ardor3d-core/src/main/java/com/ardor3d/util/geom/VertGroupData.java
@@ -21,7 +21,7 @@ public class VertGroupData {
   public static final int DEFAULT_GROUP = 0;
 
   private final Map<Long, EnumSet<MatchCondition>> _groupConditions = new HashMap<>();
-  private long[] _vertGroups = null;
+  private int[] _vertGroups = null;
 
   public VertGroupData() {}
 
@@ -33,12 +33,12 @@ public class VertGroupData {
     return _groupConditions.get(groupNumber);
   }
 
-  public long getGroupForVertex(final int index) {
+  public int getGroupForVertex(final int index) {
     if (_vertGroups != null) {
       return _vertGroups[index];
     }
     return DEFAULT_GROUP;
   }
 
-  public void setVertGroups(final long[] vertGroupMap) { _vertGroups = vertGroupMap; }
+  public void setVertGroups(final int[] vertGroupMap) { _vertGroups = vertGroupMap; }
 }
